+++ sdk/resourceconnector/arm-resourceconnector/src/operationsInterfaces/appliances.ts
@@ -12,7 +12,6 @@ import {
   ApplianceOperation,
   AppliancesListOperationsOptionalParams,
   Appliance,
-  AppliancesListBySubscriptionOptionalParams,
   AppliancesListByResourceGroupOptionalParams,
   AppliancesGetTelemetryConfigOptionalParams,
   AppliancesGetTelemetryConfigResponse,
@@ -28,7 +27,7 @@ import {
   AppliancesListKeysOptionalParams,
   AppliancesListKeysResponse,
   AppliancesGetUpgradeGraphOptionalParams,
-  AppliancesGetUpgradeGraphResponse
+  AppliancesGetUpgradeGraphResponse,
 } from "../models";
 
 /// <reference lib="esnext.asynciterable" />
@@ -39,16 +38,8 @@ export interface Appliances {
    * @param options The options parameters.
    */
   listOperations(
-    options?: AppliancesListOperationsOptionalParams
+    options?: AppliancesListOperationsOptionalParams,
   ): PagedAsyncIterableIterator<ApplianceOperation>;
-  /**
-   * Gets a list of Appliances in the specified subscription. The operation returns properties of each
-   * Appliance
-   * @param options The options parameters.
-   */
-  listBySubscription(
-    options?: AppliancesListBySubscriptionOptionalParams
-  ): PagedAsyncIterableIterator<Appliance>;
   /**
    * Gets a list of Appliances in the specified subscription and resource group. The operation returns
    * properties of each Appliance.
@@ -57,14 +48,14 @@ export interface Appliances {
    */
   listByResourceGroup(
     resourceGroupName: string,
-    options?: AppliancesListByResourceGroupOptionalParams
+    options?: AppliancesListByResourceGroupOptionalParams,
   ): PagedAsyncIterableIterator<Appliance>;
   /**
    * Gets the telemetry config.
    * @param options The options parameters.
    */
   getTelemetryConfig(
-    options?: AppliancesGetTelemetryConfigOptionalParams
+    options?: AppliancesGetTelemetryConfigOptionalParams,
   ): Promise<AppliancesGetTelemetryConfigResponse>;
   /**
    * Gets the details of an Appliance with a specified resource group and name.
@@ -75,7 +66,7 @@ export interface Appliances {
   get(
     resourceGroupName: string,
     resourceName: string,
-    options?: AppliancesGetOptionalParams
+    options?: AppliancesGetOptionalParams,
   ): Promise<AppliancesGetResponse>;
   /**
    * Creates or updates an Appliance in the specified Subscription and Resource Group.
@@ -88,7 +79,7 @@ export interface Appliances {
     resourceGroupName: string,
     resourceName: string,
     parameters: Appliance,
-    options?: AppliancesCreateOrUpdateOptionalParams
+    options?: AppliancesCreateOrUpdateOptionalParams,
   ): Promise<
     SimplePollerLike<
       OperationState<AppliancesCreateOrUpdateResponse>,
@@ -106,7 +97,7 @@ export interface Appliances {
     resourceGroupName: string,
     resourceName: string,
     parameters: Appliance,
-    options?: AppliancesCreateOrUpdateOptionalParams
+    options?: AppliancesCreateOrUpdateOptionalParams,
   ): Promise<AppliancesCreateOrUpdateResponse>;
   /**
    * Deletes an Appliance with the specified Resource Name, Resource Group, and Subscription Id.
@@ -117,7 +108,7 @@ export interface Appliances {
   beginDelete(
     resourceGroupName: string,
     resourceName: string,
-    options?: AppliancesDeleteOptionalParams
+    options?: AppliancesDeleteOptionalParams,
   ): Promise<SimplePollerLike<OperationState<void>, void>>;
   /**
    * Deletes an Appliance with the specified Resource Name, Resource Group, and Subscription Id.
@@ -128,7 +119,7 @@ export interface Appliances {
   beginDeleteAndWait(
     resourceGroupName: string,
     resourceName: string,
-    options?: AppliancesDeleteOptionalParams
+    options?: AppliancesDeleteOptionalParams,
   ): Promise<void>;
   /**
    * Updates an Appliance with the specified Resource Name in the specified Resource Group and
@@ -140,7 +131,7 @@ export interface Appliances {
   update(
     resourceGroupName: string,
     resourceName: string,
-    options?: AppliancesUpdateOptionalParams
+    options?: AppliancesUpdateOptionalParams,
   ): Promise<AppliancesUpdateResponse>;
   /**
    * Returns the cluster user credentials for the dedicated appliance.
@@ -151,7 +142,7 @@ export interface Appliances {
   listClusterUserCredential(
     resourceGroupName: string,
     resourceName: string,
-    options?: AppliancesListClusterUserCredentialOptionalParams
+    options?: AppliancesListClusterUserCredentialOptionalParams,
   ): Promise<AppliancesListClusterUserCredentialResponse>;
   /**
    * Returns the cluster customer credentials for the dedicated appliance.
@@ -162,7 +153,7 @@ export interface Appliances {
   listKeys(
     resourceGroupName: string,
     resourceName: string,
-    options?: AppliancesListKeysOptionalParams
+    options?: AppliancesListKeysOptionalParams,
   ): Promise<AppliancesListKeysResponse>;
   /**
    * Gets the upgrade graph of an Appliance with a specified resource group and name and specific release
@@ -176,6 +167,6 @@ export interface Appliances {
     resourceGroupName: string,
     resourceName: string,
     upgradeGraph: string,
-    options?: AppliancesGetUpgradeGraphOptionalParams
+    options?: AppliancesGetUpgradeGraphOptionalParams,
   ): Promise<AppliancesGetUpgradeGraphResponse>;
 }