+++ sdk/resourceconnector/arm-resourceconnector/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Release History
     
+## 2.0.0 (2024-07-15)
+    
+**Features**
+
+  - Interface AppliancesListKeysOptionalParams has a new optional parameter artifactType
+  - Enum KnownStatus has a new value EtcdSnapshotFailed
+  - Enum KnownStatus has a new value ValidatingEtcdHealth
+  - Enum KnownStatus has a new value ValidatingImageDownload
+  - Enum KnownStatus has a new value ValidatingImageUpload
+  - Enum KnownStatus has a new value ValidatingSFSConnectivity
+
+**Breaking Changes**
+
+  - Removed operation Appliances.listBySubscription
+  - Enum KnownProvider no longer has value KubeVirt
+  - Enum KnownProvider no longer has value OpenStack
+    
+    
 ## 1.0.0 (2023-08-14)
 
-The package of @azure/arm-resourceconnector is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart ).
+The package of @azure/arm-resourceconnector is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart).